Add service package in favor of micro
The micro package type aliases everything from the service package
for backwards compatibility, but is now deprecated.

The micro/test package is there to assert the micro-imported tests
work as expected.
