+++ go.mod
@@ -3,10 +3,20 @@ module github.com/nats-io/nats.go
 go 1.20
 
 require (
+	github.com/golang/protobuf v1.5.3
 	github.com/klauspost/compress v1.17.2
+	github.com/nats-io/nats-server/v2 v2.10.5
 	github.com/nats-io/nkeys v0.4.6
 	github.com/nats-io/nuid v1.0.1
-	golang.org/x/text v0.13.0
+	go.uber.org/goleak v1.3.0
+	golang.org/x/text v0.14.0
+	google.golang.org/protobuf v1.31.0
 )
 
-require golang.org/x/crypto v0.14.0 // indirect
+require (
+	github.com/minio/highwayhash v1.0.2 // indirect
+	github.com/nats-io/jwt/v2 v2.5.3 // indirect
+	golang.org/x/crypto v0.15.0 // indirect
+	golang.org/x/sys v0.14.0 // indirect
+	golang.org/x/time v0.4.0 // indirect
+)
